(PC-22628)[BSR] refactor: full circle arrow left
