+++ pro/src/pages/Offerers/Offerer/OffererDetails/OffererDetails.tsx
@@ -4,7 +4,7 @@ import { useParams } from 'react-router-dom'
 
 import { api } from 'apiClient/api'
 import { GetOffererVenueResponseModel } from 'apiClient/v1'
-import { ReactComponent as CircleArrowIcon } from 'icons/ico-circle-arrow-left.svg'
+import { ReactComponent as CircleArrowIcon } from 'icons/full-circle-arrow-left.svg'
 import { HTTP_STATUS } from 'repository/pcapi/pcapiClient'
 import { ButtonLink } from 'ui-kit'
 import { ButtonVariant } from 'ui-kit/Button/types'
